refactor: 강의 관리 페이지 리팩터링
